Config: Don't crash on falsy whitelist config

+++ lib/mix/tasks/pleroma/config.ex
@@ -249,39 +249,43 @@ defmodule Mix.Tasks.Pleroma.Config do
 
     whitelisted_configs = Pleroma.Config.get(:database_config_whitelist)
 
-    whitelisted_groups =
-      whitelisted_configs
-      |> Enum.filter(fn
-        {_group} -> true
-        _ -> false
-      end)
-      |> Enum.map(fn {group} -> group end)
-
-    whitelisted_keys =
-      whitelisted_configs
-      |> Enum.filter(fn
-        {_group, _key} -> true
-        _ -> false
-      end)
-
-    filtered =
-      from(c in ConfigDB)
-      |> Repo.all()
-      |> Enum.filter(&not_whitelisted?(&1, whitelisted_groups, whitelisted_keys))
-
-    if not Enum.empty?(filtered) do
-      shell_info("The following settings will be removed from ConfigDB:\n")
-      Enum.each(filtered, &dump(&1))
-
-      if force or shell_prompt("Are you sure you want to continue?", "n") in ~w(Yn Y y) do
-        filtered_ids = Enum.map(filtered, fn %{id: id} -> id end)
-
-        Repo.delete_all(from(c in ConfigDB, where: c.id in ^filtered_ids))
-      else
-        shell_error("No changes made.")
-      end
-    else
+    if whitelisted_configs in [nil, false] do
       shell_error("No unwanted settings in ConfigDB. No changes made.")
+    else
+      whitelisted_groups =
+        whitelisted_configs
+        |> Enum.filter(fn
+          {_group} -> true
+          _ -> false
+        end)
+        |> Enum.map(fn {group} -> group end)
+
+      whitelisted_keys =
+        whitelisted_configs
+        |> Enum.filter(fn
+          {_group, _key} -> true
+          _ -> false
+        end)
+
+      filtered =
+        from(c in ConfigDB)
+        |> Repo.all()
+        |> Enum.filter(&not_whitelisted?(&1, whitelisted_groups, whitelisted_keys))
+
+      if not Enum.empty?(filtered) do
+        shell_info("The following settings will be removed from ConfigDB:\n")
+        Enum.each(filtered, &dump(&1))
+
+        if force or shell_prompt("Are you sure you want to continue?", "n") in ~w(Yn Y y) do
+          filtered_ids = Enum.map(filtered, fn %{id: id} -> id end)
+
+          Repo.delete_all(from(c in ConfigDB, where: c.id in ^filtered_ids))
+        else
+          shell_error("No changes made.")
+        end
+      else
+        shell_error("No unwanted settings in ConfigDB. No changes made.")
+      end
     end
   end
